# api/openapi/api-gateway.yaml
openapi: 3.1.0
info:
  title: Personal Finance OS API Gateway
  version: 0.4.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/transactions:
    get:
      summary: List transactions from the ledger service
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transactions list
    post:
      summary: Store a transaction in the ledger service
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        '201':
          description: Transaction created
  /api/v1/categories:
    get:
      summary: List ledger categories
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Categories list
  /api/v1/recurring:
    get:
      summary: Detect recurring transactions from the current ledger slice
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Recurring patterns
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        user_id:
          type: string
        account_id:
          type: string
        merchant:
          type: string
        category:
          type: string
        amount_cents:
          type: integer
        currency:
          type: string
        occurred_at:
          type: string
          format: date-time